Move request function to API store for better encapsulation
